pyproject.toml
[project]
name = "imggen"
version = "0.1.0"
description = "Visual-search image generator for psychology experiments."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Bartlomiej Dudek" }]
dependencies = [
    "Pillow>=10.1",
    "numpy",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = [".venv", "__pycache__", ".pytest_cache", ".ruff_cache"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line length is enforced via the formatter only
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
